Code review comments. Replaced uint32_t with size_t. Added log to console

+++ BaikalDataGenerator/Source/config_loader.cpp
@@ -226,7 +226,7 @@ std::vector<LightInfo> ConfigLoader::Lights() const
     return m_light_settings;
 }
 
-std::vector<unsigned> ConfigLoader::Spp() const
+std::vector<size_t> ConfigLoader::Spp() const
 {
     return m_spp;
 }
